requirements.txt
accelerate==1.8.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
ccimport==0.4.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
cumm-cu120==0.4.11
cycler==0.12.1
dash==3.0.4
dataclasses-json==0.6.7
datasets==3.6.0
dearpygui==1.10.0
decorator==5.2.1
Deprecated==1.2.18
diff_gaussian_rasterization @ file:///home/liushuzhi/pax/SC-GS/submodules/diff-gaussian-rasterization
diffoctreerast @ file:///tmp/extensions/diffoctreerast
diffusers==0.34.0
dill==0.3.8
easydict==1.13
einops==0.8.1
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
flash-attn==2.7.1.post4
flash-linear-attention @ git+https://github.com/fla-org/flash-linear-attention@f26c81942dcb03899697563e224c0bd9b4c5f314
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
glcontext==3.0.0
gradio==4.44.1
gradio_client==1.3.0
gradio_litmodel3d==0.0.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
igraph==0.11.9
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
ipycanvas==0.13.3
ipyevents==2.0.2
ipython==8.37.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==7.4.9
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kaolin==0.17.0
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
moderngl==5.12.0
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.44.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==1.26.4
nvdiffrast @ file:///tmp/extensions/nvdiffrast
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.22.0
open3d==0.19.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.3.0
parso==0.8.4
pccm==0.4.16
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
plotly==6.1.2
plyfile==1.1.2
pooch==1.8.2
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pybind11==2.13.6
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyglet==2.1.6
pygltflib==1.16.4
Pygments==2.19.2
PyMatting==1.1.14
pymeshfix==0.17.1
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@e3d3a67a89907476bd5b63289f9669bd427ae550
pytz==2025.2
pyvista==0.45.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.32.4
retrying==1.4.0
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
scooby==0.10.1
segment-anything==1.0
semantic-version==2.10.0
shellingham==1.5.4
simple_knn @ file:///home/liushuzhi/pax/SC-GS/submodules/simple-knn
six==1.17.0
sniffio==1.3.1
spconv-cu120==2.3.6
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
termcolor==3.1.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.21.2
tomlkit==0.12.0
torch==2.5.0+cu121
torchaudio==2.5.0+cu121
torchvision==0.20.0+cu121
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trimesh==4.6.12
triton==3.1.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
usd-core==25.5.1
utils3d @ git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8
uvicorn==0.34.3
vtk==9.4.2
warp-lang==1.7.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wrapt==1.17.2
xatlas==0.0.10
xformers @ git+https://github.com/facebookresearch/xformers.git@526df11f09203d9191af1492e248c1df0d7c2ff1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
